requirements.txt
attrs==24.2.0
cloudpickle==3.1.0
coloredlogs==15.0.1
decorator==5.1.1
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
humanfriendly==10.0
imutils==0.5.4
Jinja2==3.1.4
MarkupSafe==3.0.2
ml_dtypes==0.5.0
mlc-ai-cpu==0.17.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
onnx==1.17.0
onnxruntime==1.20.1
onnxruntime-tools==1.7.0
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
protobuf==5.29.1
psutil==6.1.0
py-cpuinfo==9.0.0
py3nvml==0.2.7
scipy==1.14.1
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
typing_extensions==4.12.2
xmltodict==0.14.2
